requirements.txt
# All libraries for this project
fastapi==0.110.0
httpx==0.23.2
joblib==1.3.2
pandas==2.0.3
plotly==5.20.0
psycopg2==2.9.9
python-decouple==3.8
requests==2.31.0
selenium==4.17.2
tqdm==4.66.2
uvicorn==0.27.1